Move dataset/ geostore code to dedicated file
